+++ backend/internal/domain/model/user.go
@@ -2,14 +2,16 @@ package model
 
 // User represents a regular (non-admin) user in the system
 type User struct {
-	ID        string  `json:"id" bson:"_id"`
-	Email     string  `json:"email" bson:"email"`
-	Password  string  `json:"-" bson:"password"` // Never exposed in JSON
-	FirstName string  `json:"firstName" bson:"firstName"`
-	LastName  string  `json:"lastName" bson:"lastName"`
-	Address   Address `json:"address" bson:"address"`
-	Created   int64   `json:"created" bson:"created"`
-	Updated   int64   `json:"updated" bson:"updated"`
+	ID             string  `json:"id" bson:"_id"`
+	Email          string  `json:"email" bson:"email"`
+	Password       string  `json:"-" bson:"password"` // Never exposed in JSON
+	FirstName      string  `json:"firstName" bson:"firstName"`
+	LastName       string  `json:"lastName" bson:"lastName"`
+	Address        Address `json:"address" bson:"address"`
+	Created        int64   `json:"created" bson:"created"`
+	Updated        int64   `json:"updated" bson:"updated"`
+	Deactive       bool    `json:"deactive" bson:"deactive"`
+	ChangePassword bool    `json:"changePassword" bson:"changePassword"`
 }
 
 // Address represents a user's address